Improve the docker testing
Add a DockerHelper to get the url for the local docker client so you
don't have to change any code if you are using another method of
connecting to the daemon.

Also the assertions in the create docker image spec are more finegraned
and the spec only deletes the image it created.