Disable vagrant up and down
To improve the performance of the build the local docker installation
will be used. If there is no local docker installation available one
must provide a port forwarding for localhost:2375 to another host (e. g.
with ssh port forwarding).

If you have a local docker installation and it is not detected by the
build process please check the DockerHelper in the testBase module and
adapt it to your requirements.