Implement a task for kubernetes port forwarding
Port forwarding is done via the kubectl command. It is by far not
perfect but for now it should work. The open tasks creates a process and
waits till the process writes a line to stdout. The created process is
tracked via a registry.

The close task gets the process from the registry and ends it.