Generate java objects from the kubernetes API
Kubernetes provides a swagger api definition which will be used to
generate all API objects so in a later step json can be generated.

To run swagger codegen the corresponding gradle plugin is used. Because
it can't be configured enought, I copied it into the buildSrc project
and 'fixed' it.