Mark all the top level api objects with an interface
To make the gradle model api safer, mark all objects, which are allowed
to be sent to the server with the interface TopLevelApiObject.

The mustache templates aren't able to handle this complexity, so all the
classes are filtered with regex and then "fixed" with some more regix
inside the build.gradle of the api-model project.