Change the mechanism to add files to a dockerfile
Docker does only automatically extract tar files or one of their
compressed versions, but not zip files. To take adventage of this docker
feature, the files to be added are now packed into tar archives.

There are now also tests to verify that the files are actually extracted
from the archive.